fix: split amount type

+++ src/Lib/Model/Split.php
@@ -11,8 +11,8 @@ class Split
     /** @var string|null $recipientToken defines the recipient of the split */
     public ?string $recipientToken;
 
-    /** @var int|null $amount defines the fixed value that the account will receive */
-    public ?int $amount;
+    /** @var float|null $amount defines the fixed value that the account will receive */
+    public ?float $amount;
 
     /** @var float|null $percentage defines the percentage of the value that the account will receive */
     public ?float $percentage;
